Small code changes, updating module names, removed extraneous code
