+++ src/utrition/src/components/imageupload/imageupload.js
@@ -1,0 +1,49 @@
+import React, { useState } from "react";
+import axios from "axios";
+
+const ImageUpload = () => {
+  const [image, setImage] = useState("");
+  const [responseData, setResponseData] = useState("");
+  function handleImage(e) {
+    setImage(e.target.files[0].name);
+  }
+
+  function getData() {
+    axios({
+      method: "POST",
+      url: "/upload",
+      data: { path: JSON.stringify(image) },
+    })
+      .then((response) => response.json())
+      .catch((error) => console.log(error));
+
+    axios({
+      method: "GET",
+      url: "/upload",
+    })
+      .then((response) => {
+        setResponseData({
+          food: response.data,
+        });
+      })
+      .catch((error) => {
+        if (error.response) {
+          console.log(error.response);
+          console.log(error.response.status);
+          console.log(error.response.headers);
+        }
+      });
+  }
+
+  return (
+    <div style={{ flexDirection: "row" }}>
+      <input type="file" name="file" onChange={handleImage} />
+      <button onClick={getData}>Submit</button>
+      <div style={{ border: "solid" }}>
+        <pre>{JSON.stringify(responseData.food, null, 2)}</pre>
+      </div>
+    </div>
+  );
+};
+
+export default ImageUpload;
